pyproject.toml
[tool.poetry]
name = "applied-ai-rag-assistant"
version = "0.0.1"
description = "Advanced RAG assistant"
authors = ["Pierre Bittner <pierre.bittner@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "rag_assistant"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
langchain = {version = ">=0.2.0"}
openai = ">=1.6.1"
duckduckgo-search = ">4.0"
pypdf = ">=4.0.1,<5.0.0" # "^3.12.2"
sentence-transformers = "^2.2.2"
torch = ">=2.0.0, !=2.0.1"
tabulate = "^0.9.0"
streamlit-feedback = "^0.1.3"
langchain-experimental = "^0.0.59"
streamlit = ">=1.26"
docarray = "^0.40.0"
langchain-community = ">=0.2.0" # must have "^0.0.29"
langchain-openai = ">=0.1.1"
# There is a incompatibility issue with chroma and llamaindex
# was working with 0.4.3 but it is no longer supporter by llamaidnex and there is incompatibility issue
# newer version of chromadb raised "ModuleNotFoundError: No module named 'hnswlib'" in vector/local_hnsw.py", line 26
# llama-index-vector-stores-chroma (0.1.6) depends on chromadb (>=0.4.22,<0.5.0)
# at runtime
# using FAISS for now as vectorstore
# chromadb  = "0.4.3" # Must have "0.4.3"
# solved by doing pip uninstall hnswlib
# and pip install chroma-hnswlib
chromadb = ">=0.4.24"
boto3= ">=1.28.59"
chroma-hnswlib = ">=0.7.3"
numexpr = "^2.8.8"
langchainhub = "^0.1.14"
llama-index =  ">=0.10.20"
llama-index-llms-openai =  ">=0.1.12"
llama-index-readers-wikipedia = ">=0.1.3"
llama-index-vector-stores-chroma = ">=0.1.6"
llama-index-embeddings-huggingface = ">=0.1.6"
nltk = ">=3.8.1"
python-dotenv= ">=1.0.0"
PyPDF2 = ">=3.0.1"
faiss-cpu = ">=1.8.0"
wikipedia = ">=1.4.0"
pathlib = ">=1.0.1"
pytest = ">=8.1.1"
trulens_eval = ">=0.27.0"
mistralai = ">=0.1.8"
llama-index-llms-mistralai = ">=0.1.10"
llama-index-llms-bedrock = ">=0.1.8"
llama-index-llms-anthropic = ">=0.1.11"
llama-index-embeddings-mistralai = ">=0.1.4"
llama-index-embeddings-langchain = ">=0.1.2"
llama-index-embeddings-bedrock = ">=0.1.5"

# pysqlite3-binary = "^0.5.2.post3"
langchain_mistralai = ">=0.1.0"
langchain-aws = ">=0.1.4"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.4.1"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
